v0.15.0

+++ run.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/sh
 set -e
 
 echo "=> Using weave version: $VERSION"
@@ -6,32 +6,37 @@ echo "=> Using weave version: $VERSION"
 echo "=> Using docker binary:"
 docker version
 
-WEAVE_IMAGES=$(docker images | grep zettio/weave | wc -l)
-if [ ${WEAVE_IMAGES} -eq "0" ]; then
-    echo "=> Setting up weave images"
-    /weave setup
-fi
-
 if [ "${WEAVE_LAUNCH}" = "**None**" ]; then
     echo "WEAVE_LAUNCH is **None**. Not running 'weave launch'"
 else
-    echo "=> Resetting weave on the node"
-    /weave reset
+    if [ ! -f "/.weave_launched" ]; then
+        WEAVE_IMAGES=$(docker images | grep weaveworks/weave | wc -l)
+        if [ ${WEAVE_IMAGES} -eq "0" ]; then
+            echo "=> Setting up weave images"
+            /weave --local setup
+        fi
+
+        echo "=> Resetting weave on the node"
+        /weave --local reset
+    fi
+
     if [ ! -z "${WEAVE_PASSWORD}" ]; then
         echo "=> Running: weave launch -password XXXXXX ${WEAVE_LAUNCH}"
-        /weave launch -password ${WEAVE_PASSWORD} ${WEAVE_LAUNCH} || true
+        /weave --local launch -password ${WEAVE_PASSWORD} ${WEAVE_LAUNCH} || true
     else
         echo "!! WARNING: No \$WEAVE_PASSWORD set!"
         echo "=> Running: weave launch ${WEAVE_LAUNCH}"
-        /weave launch ${WEAVE_LAUNCH} || true
+        /weave --local launch ${WEAVE_LAUNCH} || true
     fi
     sleep 2
 fi
 
 echo "=> Current weave router status"
-/weave status
+/weave --local status
+
+touch /.weave_launched
 
 docker logs -f weave &
 
 echo "=> Starting peer discovery daemon"
-exec python -u /app/monitor.py $@
+exec /weave-daemon $@
